bug fix for deployment
