+++ src/lib/helpers/deploy.ts
@@ -3,7 +3,7 @@ import type { IVizKeys, IVizType, IVizValues } from '$lib/interface';
 import { viz_type, viz_keys, viz_values, color_pallette, opacity_values } from '$lib/store';
 
 function fuelStoreFromURLParams(urlParams: URLSearchParams) {
-	console.log("urlParams.get('color')", urlParams.get('color'));
+	console.log("urlParams.get('color_pallette')", urlParams.get('color_pallette'));
 	viz_type.set(urlParams.get('type') as IVizType);
 	viz_keys.set(urlParams.get('keys') as IVizKeys);
 	viz_values.set(urlParams.get('values') as IVizValues);
